### 2024-04-22 #6662 @editorialbot
[PRE REVIEW]: VisualTorch: Streamlining Visualization for PyTorch Neural Network Architectures
**Submitting author:** @willyfh (<a href="http://orcid.org/0000-0002-6209-3981">Willy Fitra Hendria</a>)
**Repository:** https://github.com/willyfh/visualtorch
**Branch with paper.md** (empty if default branch): joss-review
**Version:** v0.2.2
