### 2024-04-23 #6662 @jromanowska
In the meantime, I suggest you:
- check if you can find any DOIs for the references in the paper that our editorialbot mentioned;
- include code as code in the examples in the paper (now, it's a figure so the code cannot be copied easily).

Btw - I liked the documentation page very much!
