### 2024-04-23 #6662 @willyfh
👋 @jromanowska Thank you so much for taking charge on this.

I don't have any specific suggestions for reviewers at the moment, so I'll leave it in your capable hands.

I'll ensure to check for DOIs and update the code examples in the paper as suggested.

Glad to hear you liked the documentation page! If there's anything else I need to revise, please let me know.
