### 2024-04-26 #6678 @jromanowska
@willyfh - could you address the missing DOI issue first and re-generate the paper?
