+++ paper/paper.bib
@@ -5,7 +5,8 @@
   year={2022},
   eprint={2211.15733},
   archivePrefix={arXiv},
-  primaryClass={cs.SE}
+  primaryClass={cs.SE},
+  doi={10.48550/arxiv.2211.15733}
 }
 
 @misc{Gavrikov:2020,
@@ -25,24 +26,24 @@
   howpublished={\url{https://github.com/fchollet/keras}},
 }
 
-@incollection{Paszke:2019,
-  title = {PyTorch: An Imperative Style, High-Performance Deep Learning Library},
-  author = {Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and Desmaison, Alban and Kopf, Andreas and Yang, Edward and DeVito, Zachary and Raison, Martin and Tejani, Alykhan and Chilamkurthy, Sasank and Steiner, Benoit and Fang, Lu and Bai, Junjie and Chintala, Soumith},
-  booktitle = {Advances in Neural Information Processing Systems 32},
-  editor = {H. Wallach and H. Larochelle and A. Beygelzimer and F. d\textquotesingle Alch\'{e}-Buc and E. Fox and R. Garnett},
-  pages = {8024--8035},
-  year = {2019},
-  publisher = {Curran Associates, Inc.},
-  url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-performance-deep-learning-library.pdf}
+@misc{Paszke:2019,
+  title={PyTorch: An Imperative Style, High-Performance Deep Learning Library}, 
+  author={Adam Paszke and Sam Gross and Francisco Massa and Adam Lerer and James Bradbury and Gregory Chanan and Trevor Killeen and Zeming Lin and Natalia Gimelshein and Luca Antiga and Alban Desmaison and Andreas Köpf and Edward Yang and Zach DeVito and Martin Raison and Alykhan Tejani and Sasank Chilamkurthy and Benoit Steiner and Lu Fang and Junjie Bai and Soumith Chintala},
+  year={2019},
+  eprint={1912.01703},
+  archivePrefix={arXiv},
+  primaryClass={cs.LG},
+  doi={10.48550/arXiv.1912.01703}
 }
 
-@misc{Abadi:2015,
-  title={ {TensorFlow}: Large-Scale Machine Learning on Heterogeneous Systems},
-  url={https://www.tensorflow.org/},
-  note={Software available from tensorflow.org},
-  author={
-    Mart\'{i}n~Abadi and others},
-  year={2015},
+@misc{Abadi:2016,
+  title={TensorFlow: A system for large-scale machine learning}, 
+  author={Martín Abadi and others},
+  year={2016},
+  eprint={1605.08695},
+  archivePrefix={arXiv},
+  primaryClass={cs.DC},
+  doi={10.48550/arXiv.1605.08695}
 }
 
 @Article{Hunter:2007,
