### 2024-04-26 #6678 @willyfh
> ```
> @editorialbot commands
> ```

👋🏼 @jromanowska, I have addressed the missing DOIs first. As far as I found, for VisualKeras and Keras, there are no DOIs for the GitHub repositories. Thank you.
