### 2024-04-26 b273347
include code in the examples as `code` in the paper
