+++ paper/paper.md
@@ -37,13 +37,104 @@ One of the important features of VisualTorch is its ability to automatically map
 
 # Usage Example
 
-We provide a usage example for each layered, LeNet, and graph style visualization in \autoref{fig:layered}, \autoref{fig:lenet}, and \autoref{fig:graph}, respectively. For each example, we display the graphic generated by our library using the matplotlib library [@Hunter:2007].
+In this section, we provide a usage example for each of the layered, LeNet, and graph-style visualizations. The graphic visualizations, displayed using the Matplotlib library [@Hunter:2007], are shown in \autoref{fig:layered}, \autoref{fig:lenet}, and \autoref{fig:graph}.
 
-![An example of layered style visualization generated using VisualTorch, accompanied by its corresponding Python code.\label{fig:layered}](figure-layered.png){ width=80% }
+## Layered Style
 
-![An example of LeNet style visualization generated using VisualTorch, accompanied by its corresponding Python code.\label{fig:lenet}](figure-lenet.png){ width=80% }
+```python
+import matplotlib.pyplot as plt
+import visualtorch
+from torch import nn
 
-![An example of graph style visualization generated using VisualTorch, accompanied by its corresponding Python code.\label{fig:graph}](figure-graph.png){ width=80% }
+model = nn.Sequential(
+    nn.Conv2d(3, 16, kernel_size=3, padding=1),
+    nn.ReLU(),
+    nn.MaxPool2d(2, 2),
+    nn.Conv2d(16, 32, kernel_size=3, padding=1),
+    nn.ReLU(),
+    nn.MaxPool2d(2, 2),
+    nn.Conv2d(32, 64, kernel_size=3, padding=1),
+    nn.ReLU(),
+    nn.MaxPool2d(2, 2),
+    nn.Flatten(),
+    nn.Linear(64 * 28 * 28, 256),
+    nn.ReLU(),
+    nn.Linear(256, 10),
+)
+
+input_shape = (1, 3, 224, 224)
+
+img = visualtorch.layered_view(model, input_shape=input_shape, legend=True)
+
+plt.axis("off")
+plt.tight_layout()
+plt.imshow(img)
+plt.show()
+```
+
+![An example of layered style visualization generated using VisualTorch.\label{fig:layered}](figure-layered.png){ width=80% }
+
+## LeNet Style
+
+```python
+import matplotlib.pyplot as plt
+import visualtorch
+from torch import nn
+
+model = nn.Sequential(
+    nn.Conv2d(3, 8, kernel_size=3, padding=1),
+    nn.MaxPool2d(2, 2),
+    nn.Conv2d(8, 16, kernel_size=3, padding=1),
+    nn.MaxPool2d(2, 2),
+)
+
+input_shape = (1, 3, 128, 128)
+
+img = visualtorch.lenet_view(model, input_shape=input_shape)
+
+plt.axis("off")
+plt.tight_layout()
+plt.imshow(img)
+plt.show()
+```
+
+![An example of LeNet style visualization generated using VisualTorch.\label{fig:lenet}](figure-lenet.png){ width=80% }
+
+## Graph Style
+
+```python
+import matplotlib.pyplot as plt
+import torch
+import visualtorch
+from torch import nn
+
+class SimpleDense(nn.Module):
+    def __init__(self) -> None:
+        super().__init__()
+        self.h0 = nn.Linear(4, 8)
+        self.h1 = nn.Linear(8, 8)
+        self.h2 = nn.Linear(8, 4)
+        self.out = nn.Linear(4, 2)
+
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
+        x = self.h0(x)
+        x = self.h1(x)
[…]
