### 2024-05-27 #6678 @Asieh-A-Mofrad
Hi,
Sorry for not showing up for a while. I'm still on the task and try to do some progress during this week.
@willyfh could you please check the hyperlink on the MIT licence? The link (https://github.com/willyfh/visualtorch/blob/update-readme/LICENSE) ends up to '404 - page not found' message.
