### 2024-05-27 #6678 @Asieh-A-Mofrad
@willyfh This is the correct link: https://github.com/willyfh/visualtorch/blob/main/LICENSE
Please update the hyperlink.
