### 2024-06-01 #6678 @willyfh
> @willyfh - FYI, I've contacted the reviewers and asked them to update us on their progress.

@jromanowska  Thank you for your assistance. I will monitor for their updates and revise my submission accordingly. :+1:

> @willyfh This is the correct link: https://github.com/willyfh/visualtorch/blob/main/LICENSE Please update the hyperlink.

@Asieh-A-Mofrad Thank you for your feedback. I have fixed the hyperlink as requested :+1:
