### 2024-06-04 willyfh/visualtorch#48 @dgrell
Copyright notices from the original codebases are missing
You have marked every file with (C) 2024 Willy Fitra Hendria, also the LICENSE file only has your name in it. This is a break with the licence of the projects that you have based yours on.

For example, visualkeras (which has substantial code similarities, and which you acknowledge as a basis), has an MIT licence with "Copyright (c) 2020 Paul Gavrikov"

The MIT licence states, as its one requirement that "The above copyright notice [...] shall be included in all copies or substantial portions of the Software.
"

Action point: the LICENSE file and any code file that has substantial pieces of code from other sources, need to include the copyright line from those original sources in addition to yours.

An example for graph.py:
```
# Copyright (c) 2020 Paul Gavrikov
# Copyright (C) 2024 Willy Fitra Hendria
# SPDX-License-Identifier: MIT
```

I have not checked the other two packages you mention as inspiration, the same should be done for those authors if there is code overlap.

[https://github.com/openjournals/joss-reviews/issues/6678](url)
