### 2024-06-04 #6678 @dgrell
https://github.com/willyfh/visualtorch/issues/48 is substantial. Also, looking at the code for visualkeras, a clearer statement of which parts are new/modified in visualtorch would be helpful for the review process.
