### 2024-06-04 #6678 @dgrell
@jromanowska Question about the bullet points listed under "Documentation". Is it OK for them to be somewhere in the codebase, or is the idea that all these points should be mentioned on the documentation website https://visualtorch.readthedocs.io/ ?

I don't know what's common practice here, I personally prefer the latter.
