### 2024-06-04 willyfh/visualtorch#48 @willyfh
Hi @dgrell, thanks for your feedback. I'll thoroughly check the source code and add the appropriate copyright lines
