### 2024-06-04 willyfh/visualtorch#48 @willyfh
Hi @dgrell, thank you for your feedback. I will thoroughly check the source code and provide the appropriate copyright lines.
