### 2024-06-05 #6678 @jromanowska
> @jromanowska Question about the bullet points listed under "Documentation". Is it OK for them to be somewhere in the codebase, or is the idea that all these points should be mentioned on the documentation website [visualtorch.readthedocs.io](https://visualtorch.readthedocs.io/) ?
>
> I don't know what's common practice here, I personally prefer the latter.

All these should be easily accessible - in case of python, these should be on the documentation website :+1:
