### 2024-06-05 #6678 @jromanowska
> [willyfh/visualtorch#48](https://github.com/willyfh/visualtorch/issues/48) is substantial. Also, looking at the code for visualkeras, a clearer statement of which parts are new/modified in visualtorch would be helpful for the review process.

Thank you for checking! We'll wait for @willyfh to explain in details which parts are new or modified - this is an important issue, as you noticed. As long as _your_ code provides new functionality, which is useful for researchers, using code from another software is fine, however, it needs to be clearly stated and properly cited, using correct licences, etc.
