### 2024-06-06 #6678 @Asieh-A-Mofrad
@jromanowska I'm almost done with the review process.
The paper appears to be well written and situate VisualTorch within the field of neural network visualization tools. There are some repetitions but it is structured logically, with sections for the summary, statement of need, introduction, usage examples, and acknowledgements. @willyfh Please check the reference to TensorFlow (Abadi:2015?) and edit it.
