+++ paper/paper.md
@@ -29,7 +29,7 @@ In deep learning projects based on Keras [@Chollet:2015], the visualkeras projec
 
 # Introduction
 
-Recent advancements in artificial intelligence have sparked widespread interest among researchers, particularly in exploring innovative algorithmic approaches such as neural networks or deep learning architectures. These architectures have demonstrated remarkable utility across various AI applications, including computer vision, natural language processing, and robotics. For implementing the neural network architectures, many researchers and practitioners often utilize established deep learning frameworks, such as PyTorch [@Paszke:2019], TensorFlow [@Abadi:2015], and Keras [@Chollet:2015]. Among the existing deep learning frameworks, PyTorch has been gaining increasing popularity, experiencing a surge in adoption within the AI community in recent years [@Aoun:2022].
+Recent advancements in artificial intelligence have sparked widespread interest among researchers, particularly in exploring innovative algorithmic approaches such as neural networks or deep learning architectures. These architectures have demonstrated remarkable utility across various AI applications, including computer vision, natural language processing, and robotics. For implementing the neural network architectures, many researchers and practitioners often utilize established deep learning frameworks, such as PyTorch [@Paszke:2019], TensorFlow [@Abadi:2016], and Keras [@Chollet:2015]. Among the existing deep learning frameworks, PyTorch has been gaining increasing popularity, experiencing a surge in adoption within the AI community in recent years [@Aoun:2022].
 
 To effectively communicate their ideas, researchers often employ architecture diagrams as aids for comprehension. While detailed mathematical descriptions help in understanding the intricacies of algorithms, visual representations of architectures offer an additional means of conveying concepts, enabling individuals to grasp the overall visual representation. Our VisualTorch library is designed to streamline the visualization of PyTorch-based neural network architectures. Instead of manually constructing diagrams from scratch, researchers can leverage a single function call of our library to generate visualizations. With a variety of customization options, users can tailor visualizations to suit their preferences.
 
