### 2024-06-25 f381bab
add the list of dependencies in the docs
