+++ docs/source/snippets/install/requirements.txt
@@ -1,0 +1,4 @@
+pillow>=10.0.0
+numpy>=1.18.1
+aggdraw>=1.3.11
+torch>=2.0.0
