### 2024-06-25 097b63a
Update documentation based on the bullet points of JOSS review checklist (#52)
* add the list of dependencies in the docs

* add developer guides to the docs
