+++ visualtorch/utils/layer_utils.py
@@ -1,5 +1,6 @@
 """Layer Utils module for pytorch model visualization."""
 
+# Copyright (C) 2020 Paul Gavrikov
 # Copyright (C) 2024 Willy Fitra Hendria
 # SPDX-License-Identifier: MIT
 
