### 2024-06-25 d5cd146
Reduce repetition and clarify newly added features in VisualTorch
