+++ paper/paper.md
@@ -29,11 +29,11 @@ In deep learning projects based on Keras [@Chollet:2015], the visualkeras projec
 
 # Introduction
 
-Recent advancements in artificial intelligence have sparked widespread interest among researchers, particularly in exploring innovative algorithmic approaches such as neural networks or deep learning architectures. These architectures have demonstrated remarkable utility across various AI applications, including computer vision, natural language processing, and robotics. For implementing the neural network architectures, many researchers and practitioners often utilize established deep learning frameworks, such as PyTorch [@Paszke:2019], TensorFlow [@Abadi:2016], and Keras [@Chollet:2015]. Among the existing deep learning frameworks, PyTorch has been gaining increasing popularity, experiencing a surge in adoption within the AI community in recent years [@Aoun:2022].
+Recent advancements in artificial intelligence have sparked widespread interest among researchers, particularly in exploring innovative algorithmic approaches such as neural networks or deep learning architectures. These architectures have demonstrated remarkable utility across various AI applications, including computer vision, natural language processing, and robotics. For implementing the neural network architectures, many researchers and practitioners often utilize established deep learning frameworks, such as PyTorch [@Paszke:2019], TensorFlow [@Abadi:2016], and Keras [@Chollet:2015]
 
-To effectively communicate their ideas, researchers often employ architecture diagrams as aids for comprehension. While detailed mathematical descriptions help in understanding the intricacies of algorithms, visual representations of architectures offer an additional means of conveying concepts, enabling individuals to grasp the overall visual representation. Our VisualTorch library is designed to streamline the visualization of PyTorch-based neural network architectures. Instead of manually constructing diagrams from scratch, researchers can leverage a single function call of our library to generate visualizations. With a variety of customization options, users can tailor visualizations to suit their preferences.
+To effectively communicate their ideas, researchers often employ architecture diagrams as aids for comprehension. While detailed mathematical descriptions help in understanding the intricacies of algorithms, visual representations of architectures offer an additional means of conveying concepts, enabling individuals to grasp the overall visual representation. Our VisualTorch library is designed to facilitate the visualization of PyTorch-based neural network architectures. Instead of manually constructing diagrams from scratch, researchers can simply leverage our library to generate visualizations. With a variety of customization options, users can tailor visualizations to suit their preferences.
 
-One of the important features of VisualTorch is its ability to automatically map a neural network model to visualizations using various styles such as layered, graph, and LeNet-like visualization [@Lecun:1998]. Users can further refine these visualizations by adjusting attributes such as color, opacity, and size. VisualTorch aims to offer a solution for rapidly visualizing a wide range of neural network architectures in PyTorch. Inspired by the visualkeras [@Gavrikov:2020] project, our VisualTorch library shares a similar motivation but offers enhanced functionality specifically tailored for PyTorch-based architectures. In addition to providing more visualization styles, our library also provides online web-based documentation and streamlined CI/CD workflows, which improve usability and facilitate future development.
[…]
