### 2024-06-26 #6678 @willyfh
Thank you for your patience in waiting for my responses. I appreciate the valuable feedback and suggestions from the reviewers and editor. @jromanowska @Asieh-A-Mofrad @dgrell

Below are my responses:

# Reviewer: @Asieh-A-Mofrad

> There are some repetitions but it is structured logically, with sections for the summary, statement of need, introduction, usage examples, and acknowledgements.

@Asieh-A-Mofrad Thank you for your comment. To reduce repetition, I removed a repetitive sentence and rephrased some words in the revised manuscript as follows:

- **(Introduction, Paragraph 1):** <strike>Among the existing deep learning frameworks, PyTorch has been gaining increasing popularity, experiencing a surge in adoption within the AI community in recent years [@Aoun:2022].</strike>
- **(Introduction, Paragraph 2)**: Our VisualTorch library is designed to <strike>streamline</strike> facilitate the visualization of PyTorch-based neural network architectures. Instead of manually constructing diagrams from scratch, researchers can <strike>leverage a single function call of</strike> simply leverage our library to generate visualizations. With a variety of customization options, users can tailor visualizations to suit their preferences.

> Please check the reference to TensorFlow (Abadi:2015?) and edit it.

@Asieh-A-Mofrad I have also fixed this issue in the first paragraph of the Introduction in the revised manuscript. Thank you for your feedback.

# Reviewer: @dgrell

> willyfh/visualtorch#48 is substantial

@dgrell Thank you for your feedback. Regarding the copyright issue, I have added the copyright of "Copyright (C) 2020 Paul Gavrikov" to the files in which include the borrowed code from visualkeras. Those files are as follows:

- https://github.com/willyfh/visualtorch/blob/main/visualtorch/graph.py
- https://github.com/willyfh/visualtorch/blob/main/visualtorch/utils/utils.py
- https://github.com/willyfh/visualtorch/blob/main/visualtorch/utils/layer_utils.py
- https://github.com/willyfh/visualtorch/blob/main/visualtorch/layered.py

For the other files, the content was mainly added specifically for VisualTorch, such as files for lenet-style visualization, online documentation, workflows for CI/CD, and other configuration files (.yaml).

For the two other libraries mentioned as our inspiration, i.e., pytorch-summary and torchviz, I did not borrow the exact code as I did from visualkeras. Hence, perhaps we don't need to mention the copyright for those libraries?

Specifically, I was inspired by pytorch-summary in how they utilize the `register_hook` function of the PyTorch library to collect the module and output shape. However, our implementation is different, as can be compared between my `register_hook` implementation (https://github.com/willyfh/visualtorch/blob/main/visualtorch/utils/layer_utils.py#L159) and the torchsummary implementation (https://github.com/sksq96/pytorch-summary/blob/master/torchsummary/torchsummary.py#L23C9-L23C22).

For torchviz, I was inspired by the logic of `add_base_tensor` (https://github.com/szagoruyko/pytorchviz/blob/master/torchviz/dot.py#L146) and `add_nodes` (https://github.com/szagoruyko/pytorchviz/blob/master/torchviz/dot.py#L96) to extract the graph structure in the library. In VisualTorch, I adapted the idea with different detailed implementations of the functions, as can be checked in function `_add_base_tensor` (https://github.com/willyfh/visualtorch/blob/main/visualtorch/utils/layer_utils.py#L228) and `_add_nodes` (https://github.com/willyfh/visualtorch/blob/main/visualtorch/utils/layer_utils.py#L196).

However, I am not fully sure if I should also add the copyright for such cases or not. I appreciate your feedback on this.

> Also, looking at the code for visualkeras, a clearer statement of which parts are new/modified in visualtorch would be helpful for the review process.

@dgrell Thank you for your comments. In comparison with visualkeras, I have provided functionality to visualize PyTorch-based models defined with `torch.nn.Module` and `torch.nn.Sequential`. I believe this functionality is useful for researchers, considering how popular the PyTorch framework is in the research community recently. In addition, I provide a new LeNet-like visualization, which is one of popular styles in research papers to illustrate neural network models, specifically CNNs. For better usage and development, I further provide better documentation, CI/CD, and testing.

Following are the summary of the new or modified parts:

- New -> Visualization of torch-based models for both layered style and graph style, including models defined by `torch.nn.Sequential` and custom models inherited from `torch.nn.Module`.
- New -> LeNet-like visualization.
- New -> Support for custom transparency of the colors.
- New -> Online documentation.
- New -> Workflow CI/CD, including linter, prettier, ruff, pre-commit, and pytest with multiple python versions on pull request.
- Modified -> Some other minor modifications such as docstring upgrades, code structures, bug fixing, etc.

I have rephrased the third paragraph of the introduction in the revised manuscript as follows:

"One of the important features of VisualTorch is its ability to automatically map a neural network model to visualizations using various styles such as layered, graph, and LeNet-like visualization [@Lecun:1998]. Users can further refine these visualizations by adjusting attributes such as color, opacity, and size. VisualTorch aims to offer a solution for rapidly visualizing a wide range of neural network architectures in PyTorch. Inspired by the visualkeras [@Gavrikov:2020] project, our VisualTorch library shares a similar motivation: to assist in visualizing neural network architectures. Unlike visualkeras, our library offers enhanced functionality specifically tailored for PyTorch-based architectures, supporting models defined with `torch.nn.Module` and `torch.nn.Sequential`. In addition to providing more visualization styles, including the recent addition of LeNet style, our library also provides online web-based documentation and streamlined CI/CD workflows, which improve usability and facilitate future development."

> Question about the bullet points listed under "Documentation". Is it OK for them to be somewhere in the codebase, or is the idea that all these points should be mentioned on the documentation website https://visualtorch.readthedocs.io/ ? I don't know what's common practice here, I personally prefer the latter.

@dgrell Thank you for your feedback. I have added the points listed under "Documentation" to the website, in this PR (https://github.com/willyfh/visualtorch/pull/52). In summary, I have added the list of dependencies in the `Installation` page of the online documentation, and I have also added some guidelines in the `Contributing to VisualTorch` page, which includes how to contribute, how to report issues, how to seek support, and steps on how to test the library using pre-commit and pytest.

CC: @jromanowska
