+++ paper/paper.bib
@@ -27,7 +27,7 @@
 }
 
 @misc{Paszke:2019,
-  title={PyTorch: An Imperative Style, High-Performance Deep Learning Library}, 
+  title={PyTorch: An Imperative Style, High-Performance Deep Learning Library},
   author={Adam Paszke and Sam Gross and Francisco Massa and Adam Lerer and James Bradbury and Gregory Chanan and Trevor Killeen and Zeming Lin and Natalia Gimelshein and Luca Antiga and Alban Desmaison and Andreas Köpf and Edward Yang and Zach DeVito and Martin Raison and Alykhan Tejani and Sasank Chilamkurthy and Benoit Steiner and Lu Fang and Junjie Bai and Soumith Chintala},
   year={2019},
   eprint={1912.01703},
@@ -37,7 +37,7 @@
 }
 
 @misc{Abadi:2016,
-  title={TensorFlow: A system for large-scale machine learning}, 
+  title={TensorFlow: A system for large-scale machine learning},
   author={Martín Abadi and others},
   year={2016},
   eprint={1605.08695},
