### 2024-07-27 c3174f5
Merge JOSS paper review branch (#54)
* add paper for submission to joss

* Fix missing DOIs

* include code in the examples as `code` in the paper

* Fix link to the license (#46) (#47)

* update reference to tensorflow

* Reduce repetition and clarify newly added features in VisualTorch

* revise the paper v2

* update by pre-commit
