+++ paper/paper.md
@@ -1,0 +1,143 @@
+---
+title: "VisualTorch: Streamlining Visualization for PyTorch Neural Network Architectures"
+tags:
+  - Python
+  - PyTorch
+  - Visualization
+  - Neural Networks
+  - Artificial Intelligence
+
+authors:
+  - name: Willy Fitra Hendria
+    affiliation: 1
+affiliations:
+  - name: Independent Researcher
+    index: 1
+date: 18 February 2024
+bibliography: paper.bib
+---
+
+# Summary
+
+VisualTorch is a library designed for visualizing neural network architectures in PyTorch. It offers support for multiple visualization styles, such as layered-style, graph-style, and the newly added LeNet-like visualization. When provided with a sequential or custom PyTorch model, alongside the input shape and visualization specifications, VisualTorch automatically translates the model structure into an architectural diagram. The resulting diagram can be refined using various configurations, including style, color, opacity, size, and a legend. VisualTorch is particularly valuable for projects involving PyTorch-based neural networks. By facilitating the generation of graphics with a single function call, it streamlines the process of visualizing neural network architectures. This ensures that the produced results are suitable for publication with minimal additional modifications. Moreover, owing to its diverse customization options, VisualTorch empowers users to generate polished figures suitable for publication.
+
+# Statement of Need
+
+Neural network architecture visualization plays an important role in the scientific process within the realm of artificial intelligence and machine learning. While mathematical equations and descriptive paragraphs provide detailed information about architectures, effective visualizations can significantly aid scientists in communicating their proposed architectures to others.
+
+In deep learning projects based on Keras [@Chollet:2015], the visualkeras project [@Gavrikov:2020] has been gaining traction over nearly four years of development. It offers visualization of Keras-based neural network architectures in two styles: layered and graph. The visualtorch library provides visualization for PyTorch-based architectures [@Paszke:2019]. PyTorch itself has gained popularity among deep learning frameworks within the scientific community in recent years [@Aoun:2022]. Our VisualTorch library offers more visualization styles, improved usability, and a development environment by providing web-based documentation and CI/CD pipelines for seamless future contributions and collaborations.
+
+# Introduction
+
+Recent advancements in artificial intelligence have sparked widespread interest among researchers, particularly in exploring innovative algorithmic approaches such as neural networks or deep learning architectures. These architectures have demonstrated remarkable utility across various AI applications, including computer vision, natural language processing, and robotics. For implementing the neural network architectures, many researchers and practitioners often utilize established deep learning frameworks, such as PyTorch [@Paszke:2019], TensorFlow [@Abadi:2016], and Keras [@Chollet:2015]
+
[…]
