+++ paper/paper.bib
@@ -1,0 +1,73 @@
+
+ @misc{Aoun:2022,
+  title={An Empirical Study of Library Usage and Dependency in Deep Learning Frameworks},
+  author={Mohamed Raed El Aoun and Lionel Nganyewou Tidjon and Ben Rombaut and Foutse Khomh and Ahmed E. Hassan},
+  year={2022},
+  eprint={2211.15733},
+  archivePrefix={arXiv},
+  primaryClass={cs.SE},
+  doi={10.48550/arxiv.2211.15733}
+}
+
+@misc{Gavrikov:2020,
+  author = {Gavrikov, Paul},
+  title = {visualkeras},
+  year = {2020},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/paulgavrikov/visualkeras}},
+}
+
+@misc{Chollet:2015,
+  title={Keras},
+  author={Chollet, Fran\c{c}ois and others},
+  year={2015},
+  publisher={GitHub},
+  howpublished={\url{https://github.com/fchollet/keras}},
+}
+
+@misc{Paszke:2019,
+  title={PyTorch: An Imperative Style, High-Performance Deep Learning Library},
+  author={Adam Paszke and Sam Gross and Francisco Massa and Adam Lerer and James Bradbury and Gregory Chanan and Trevor Killeen and Zeming Lin and Natalia Gimelshein and Luca Antiga and Alban Desmaison and Andreas Köpf and Edward Yang and Zach DeVito and Martin Raison and Alykhan Tejani and Sasank Chilamkurthy and Benoit Steiner and Lu Fang and Junjie Bai and Soumith Chintala},
+  year={2019},
+  eprint={1912.01703},
+  archivePrefix={arXiv},
+  primaryClass={cs.LG},
+  doi={10.48550/arXiv.1912.01703}
+}
+
+@misc{Abadi:2016,
+  title={TensorFlow: A system for large-scale machine learning},
+  author={Martín Abadi and others},
+  year={2016},
+  eprint={1605.08695},
+  archivePrefix={arXiv},
+  primaryClass={cs.DC},
+  doi={10.48550/arXiv.1605.08695}
+}
+
+@Article{Hunter:2007,
+  Author    = {Hunter, J. D.},
+  Title     = {Matplotlib: A 2D graphics environment},
+  Journal   = {Computing in Science \& Engineering},
+  Volume    = {9},
+  Number    = {3},
+  Pages     = {90--95},
+  abstract  = {Matplotlib is a 2D graphics package used for Python for
+  application development, interactive scripting, and publication-quality
+  image generation across user interfaces and operating systems.},
+  publisher = {IEEE COMPUTER SOC},
+  doi       = {10.1109/MCSE.2007.55},
+  year      = 2007
+}
+
+@ARTICLE{Lecun:1998,
+  author={Lecun, Y. and Bottou, L. and Bengio, Y. and Haffner, P.},
+  journal={Proceedings of the IEEE},
+  title={Gradient-based learning applied to document recognition},
+  year={1998},
+  volume={86},
+  number={11},
+  pages={2278-2324},
+  keywords={Neural networks;Pattern recognition;Machine learning;Optical character recognition software;Character recognition;Feature extraction;Multi-layer neural network;Optical computing;Hidden Markov models;Principal component analysis},
+  doi={10.1109/5.726791}}
