### 2024-07-31 #6678 @jromanowska
Dear all! I'm just back from summer holidays and wanted to check how are things. @dgrell , @Asieh-A-Mofrad - how are you doing?
