### 2024-08-26 #6678 @dgrell
It's not easy for me to come to a conclusion here, and I have been going back and forth on several issues around good free software and research publication customs.

1) Looking at the commit history, visualtorch looks like it started as a fork of visualkeras, with modifications and adaptations, of which the most
substantial is the new LeNet visualization (ca. 330 lines).

The fork is not explicitly done in github. Instead, the files have been copied over into a new empty repo. This loses information about the relation between the projects. The relation is mentioned as "inspired by" in the paper, where "based on" or "adapted from" may be a better description.

2) I'm inclined to only consider LeNet as substantially new, the other modifications such as improved documentation etc are quality of use improvements which do not change the fact that the core workflow is based off visualkeras. It would have been nice to feed these improvements such as test cases etc. back to the original visualkeras project, in the spirit of the CONTRIBUTING document.

3) Now that the copyright notices have been preserved (COPYING is missing the line still), the MIT licence of visualkeras certainly permits the reuse of the code in the visualtorch repo. The repo as such is fine with the above mentioned caveats.

However, as to research publication ethics, I am reluctant to create a single-author paper here for a work that is substantially someone else's in core idea and large parts of the implementation. To me a joint authorship here seems more appropriate.

This last point creates a question to the editors, in how much this journal will grant papers to downstream forks of repos, when they layer some improvement on top of a substantial pre-existing older work.
