### 2024-08-26 #6678 @willyfh
Hi @dgrell , thank you for your comments. I just want to comment a bit about this part

> I'm inclined to only consider LeNet as substantially new, the other modifications such as improved documentation etc are quality of use improvements which do not change the fact that the core workflow is based off visualkeras. It would have been nice to feed these improvements such as test cases etc. back to the original visualkeras project, in the spirit of the CONTRIBUTING document.

In addition to lenet, documentation, and test case as you mentioned above, i think the main modification is also on the visualizing PyTorch-based models. Different than Keras, PyTorch framework has its own structure. As visualkeras specifically developed only for Keras, ans they doesn't support PyTorch as is. It requires specific modification tailored for PyTorch framework. I agree making documentation improvement and test case improvement to the visualkeras can be good contributions. However, the documentation and the test case that I made specifically used for PyTorch based model, as you can see from all examples in the documentation and the test cases are using PyTorch models. So it requires some changes, at the minimum (maybe more) on the PyTorch model parts, before I can fully integrate the documentation and test case.

Thank you again for your feedbacks. I'll leave the decision to the reviewers and editors. ☺️
