### 2024-08-26 #6678 @dgrell
Hi @willyfh , my apologies! Of course the adaptation of the code to take input from the torch data model rather than the keras one is a new step, too, and I meant to list that above together with the LeNet point.  Taken together, in my view we're still right around the boundary for substantiality which the journal instructions mention.

From my OSS tool builder's perspective, it would be cool if your two projects could actually combine into one tool that can take input from both the different libraries, since the imaging code has so much in common.
