### 2024-08-27 #6678 @jromanowska
I think the points described by @dgrell are very important here:
- every re-use of code needs a proper citation and it needs to be visible that the code was re-used
- every extension of existing software needs to be substantial enough to be accepted for publishing in JOSS

I propose that @willyfh chooses to either:
- invite the author(s) of visualkeras to this publication,
or
- re-organize the code to explicitly fork the other repo(s) so that it's easier to understand what's new here.

After that, we will need to decide whether to continue with this review process or perhaps start a new submission.
Would that be OK, @dgrell , @Asieh-A-Mofrad , @willyfh ?
