### 2024-08-27 #6678 @willyfh
@jromanowska @dgrell @Asieh-A-Mofrad Ah, by the way, if I go with the first option, would it be appropriate to invite Paul Gavrikov as a co-author of the paper? He seems to be the primary contributor to the project, especially considering the code that I’ve adopted. Another significant contributor, Soontosh, only recently contributed to the library about two months ago, and it seems that his work on VisualKeras are not incorporated into VisualTorch.
