### 2024-08-28 #6678 @willyfh
> Yes, that would be a great option. Depending on the interest, the paper could then cover both tools?

@dgrell Not sure if it is possible to cover both tools. In my understanding, one paper should be associated with one repo / tool?  cc: @jromanowska

Also, the journal recommend the number of words are within 250-1000. Even with only visualtorch, it's already almost 1000 words 😢
