### 2024-08-28 #6678 @jromanowska
That's true - the paper should be max.1000 words and focus on one tool. Sometimes, we get submissions about tool collections, though - this could be OK but then, I'd have to recommend re-submission since the topic will be broader anyway.
Let's see whether you can strengthen your submission, @willyfh , by inviting some co-authors :)
