### 2024-08-28 #6678 @willyfh
> That's true - the paper should be max.1000 words and focus on one tool. Sometimes, we get submissions about tool collections, though - this could be OK but then, I'd have to recommend re-submission since the topic will be broader anyway. Let's see whether you can strengthen your submission, @willyfh , by inviting some co-authors :)

Great! Paul Gavrikov recently replied to my email and is willing to become a co-author for this paper. Could you advise me on the procedure to add him as a co-author? Is it sufficient to simply list him in the paper, or is there anything he needs to do from his end? 🙏
