### 2024-08-28 #6678 @jromanowska
Wonderful! He needs to read the paper and perhaps you should give him some time to comment on the content of both, the paper, and the code. You can also point him to this review issue. He doesn't need to be active here, but he's welcome to give any input, if he wishes.
