### 2024-09-05 #6678 @paulgavrikov
Hi all! I am Paul Gavrikov, maintainer of VisualKeras - thank you so much for considering me as co-author! @willyfh reached out to me and I am more than happy to co-author this paper. I have sent him some minor comments about the article but otherwise it looks good to me!
