### 2024-09-05 #6678 @jromanowska
Thank you, @paulgavrikov !
@willyfh , please go ahead and create a new PDF when you're done with changes to the paper.
I will then need the reviewers, @dgrell and @Asieh-A-Mofrad, to read it once more and comment here if they recommend any more changes or if they are satisfied with the work that has been done up to now.
Also, @dgrell , if you could go through your checklist and tick off all the points that have been resolved.
