+++ paper/paper.md
@@ -10,9 +10,16 @@ tags:
 authors:
   - name: Willy Fitra Hendria
     affiliation: 1
+  - name: Paul Gavrikov
+    affiliation: [2, 3]
 affiliations:
   - name: Independent Researcher
     index: 1
+  - name: Offenburg University, 
+    index: 2
+  - name: University of Mannheim
+    index: 3
+    
 date: 18 February 2024
 bibliography: paper.bib
 ---
