### 2024-09-22 b0fa937
Update paper based on feedback from Paul Gavrikov
