+++ paper/paper.md
@@ -32,13 +32,13 @@ VisualTorch is a library designed for visualizing neural network architectures i
 
 Neural network architecture visualization plays an important role in the scientific process within the realm of artificial intelligence and machine learning. While mathematical equations and descriptive paragraphs provide detailed information about architectures, effective visualizations can significantly aid scientists in communicating their proposed architectures to others.
 
-In deep learning projects based on Keras [@Chollet:2015], the visualkeras project [@Gavrikov:2020] has been gaining traction over nearly four years of development. It offers visualization of Keras-based neural network architectures in two styles: layered and graph. The visualtorch library provides visualization for PyTorch-based architectures [@Paszke:2019]. PyTorch itself has gained popularity among deep learning frameworks within the scientific community in recent years [@Aoun:2022]. Our VisualTorch library offers more visualization styles, improved usability, and a development environment by providing web-based documentation and CI/CD pipelines for seamless future contributions and collaborations.
+In deep learning projects based on Keras [@Chollet:2015], the visualkeras project [@Gavrikov:2020] has been gaining traction over nearly four years of development. It offers visualization of Keras-based neural network architectures in two styles: layered and graph. The visualtorch library provides visualization for PyTorch-based architectures [@Paszke:2019]. PyTorch itself has gained popularity among deep learning frameworks within the scientific community in recent years [@Aoun:2022]. VisualTorch expands the functionality of VisualKeras to PyTorch and offers more visualization styles, improved usability, and a development environment by providing web-based documentation and CI/CD pipelines for seamless future contributions and collaborations.
 
 # Introduction
 
-Recent advancements in artificial intelligence have sparked widespread interest among researchers, particularly in exploring innovative algorithmic approaches such as neural networks or deep learning architectures. These architectures have demonstrated remarkable utility across various AI applications, including computer vision, natural language processing, and robotics. For implementing the neural network architectures, many researchers and practitioners often utilize established deep learning frameworks, such as PyTorch [@Paszke:2019], TensorFlow [@Abadi:2016], and Keras [@Chollet:2015]
+Recent advancements in artificial intelligence have sparked widespread interest among researchers, particularly in exploring innovative algorithmic approaches such as neural networks or deep learning architectures. These architectures have demonstrated remarkable utility across various AI applications, including computer vision, natural language processing, and robotics. To implement neural network architectures, many researchers and practitioners often utilize established deep learning frameworks, such as PyTorch [@Paszke:2019], TensorFlow [@Abadi:2016], and Keras [@Chollet:2015]
 
-To effectively communicate their ideas, researchers often employ architecture diagrams as aids for comprehension. While detailed mathematical descriptions help in understanding the intricacies of algorithms, visual representations of architectures offer an additional means of conveying concepts, enabling individuals to grasp the overall visual representation. Our VisualTorch library is designed to facilitate the visualization of PyTorch-based neural network architectures. Instead of manually constructing diagrams from scratch, researchers can simply leverage our library to generate visualizations. With a variety of customization options, users can tailor visualizations to suit their preferences.
+To effectively communicate their ideas, practitioners often employ architecture diagrams as aids for comprehension. While detailed mathematical descriptions help in understanding the intricacies of algorithms, visual representations of architectures offer an additional means of conveying concepts, enabling individuals to grasp the overall visual representation. VisualTorch is designed to facilitate the visualization of PyTorch-based neural network architectures. Instead of manually constructing diagrams from scratch, practitioners can simply leverage our library to generate visualizations. With a variety of customization options, users can tailor visualizations to suit their preferences.
 
[…]
