+++ paper/paper.md
@@ -9,14 +9,16 @@ tags:
 
 authors:
   - name: Willy Fitra Hendria
-    affiliation: '1'
+    affiliation: "1"
   - name: Paul Gavrikov
-    affiliation: '2'
+    affiliation: "2, 3"
 affiliations:
   - index: 1
     name: Independent Researcher
   - index: 2
-    name: Offenburg University and University of Mannheim
+    name: Offenburg University
+  - index: 3
+    name: University of Mannheim
     
     
 date: 18 February 2024
