### 2024-09-22 #6678 @willyfh
Thanks to @paulgavrikov, the multiple affiliations finally work :)
