### 2024-09-22 #6678 @willyfh
Hi @jromanowska, @dgrell, @Asieh-A-Mofrad,

Thank you for your patience. I've created a new PDF (above), which now includes @paulgavrikov as a co-author, and I’ve also updated the paper based on his feedback. You can review the changes in this PR: https://github.com/willyfh/visualtorch/pull/58

Thank you for your continued support. 🙏
