### 2024-09-28 a065433
Add ORCID and update some sentences (#58)
* Merge from Main (#56)

* Merge JOSS paper review branch (#54)

* add paper for submission to joss

* Fix missing DOIs

* include code in the examples as `code` in the paper

* Fix link to the license (#46) (#47)

* update reference to tensorflow

* Reduce repetition and clarify newly added features in VisualTorch

* revise the paper v2

* update by pre-commit

* Update LICENSE (#55)

* Add a co-author

* Update paper based on feedback from Paul Gavrikov

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* 
[…]
