### 2024-09-28 #6678 @willyfh
Hi @jromanowska,

Thank you so much for your patience! we’ve completed the additional tasks, and you can find all the details below.
Please let us know if we need to do anything else.

---

**✅ Double check authors and affiliations (including ORCIDs)**
We have double-checked and added the ORCIDs.

**✅ Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.**
    v0.2.3 (https://github.com/willyfh/visualtorch/releases/tag/v0.2.3)

**✅ Archive the release on Zenodo/figshare/etc and post the DOI here.**
https://doi.org/10.5281/zenodo.13852642

**✅ Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.**
The title and author list in Zenodo match those in the paper.

**✅ Make sure that the license listed for the archive is the same as the software license.**
Both Zenodo and the software have an MIT License.
