### 2024-09-30 #6678 @willyfh
> ⚠️ [Error preparing paper acceptance](https://github.com/openjournals/joss-papers/actions/runs/11101518080?check_suite_focus=true).

@jromanowska maybe this is because i released the final version using main branch?
