### 2024-09-30 #6678 @jromanowska
> > ⚠️ [Error preparing paper acceptance](https://github.com/openjournals/joss-papers/actions/runs/11101518080?check_suite_focus=true).
>
> @jromanowska maybe this is because i released the final version using main branch?

I think this should be OK, but your paper-containing branch should not be deleted... Trying again.
