### 2024-10-07 #6678 @crvernon
:wave: @willyfh - I just need you to address the following before I move forward with accepting this for publication:

**In the paper:**
- [ ] LINE 38:  Missing period at the end of the sentence.

That's it!  Great job giving me a clean submission!
