### 2024-10-08 #6678 @willyfh
> LINE 38: Missing period at the end of the sentence.

@crvernon Done 🙏 Thank you for the feedback!
